rename CRC functions to make it more intuitive

+++ one-third-hal/core/math/crc.h
@@ -4,8 +4,8 @@
 
 // ============================================================================
 typedef struct {
-    uint16_t (*_16bit8)(uint8_t*, uint16_t, uint8_t);
-    uint32_t (*_32bit8)(uint8_t*, uint16_t);
+    uint16_t (*_16bit8)(uint8_t*, uint32_t, uint8_t);
+    uint32_t (*_32bit8)(uint8_t*, uint32_t);
 } CrcSoftApi_t;
 
 typedef struct {
